refactor: add generic where specification provider impl
