+++ src/main/java/br/com/thiagoRDS/api_authors/providers/GenericWhereSpecificationProvider/GerericWhereSpecificationProvider.java
@@ -1,0 +1,9 @@
+package br.com.thiagoRDS.api_authors.providers.GenericWhereSpecificationProvider;
+
+import org.springframework.data.jpa.domain.Specification;
+
+public abstract class GerericWhereSpecificationProvider<G> {
+  public abstract Specification<G> like(String filter, String fieldPath);
+
+  public abstract Specification<G> notNull(String fieldPath);
+}
